refactor: `Fixture` 패키지 이동
