+++ src/test/java/roomescape/application/reservation/ReservationLookupServiceTest.java
@@ -14,9 +14,7 @@ import roomescape.application.ServiceTest;
 import roomescape.application.reservation.dto.request.ReservationRequest;
 import roomescape.application.reservation.dto.response.ReservationResponse;
 import roomescape.application.reservation.dto.response.ReservationStatusResponse;
-import roomescape.application.reservation.fixture.ReservationFixture;
 import roomescape.domain.member.Member;
-import roomescape.domain.member.MemberFixture;
 import roomescape.domain.member.MemberRepository;
 import roomescape.domain.reservation.BookStatus;
 import roomescape.domain.reservation.Reservation;
@@ -26,6 +24,8 @@ import roomescape.domain.reservation.ReservationTime;
 import roomescape.domain.reservation.ReservationTimeRepository;
 import roomescape.domain.reservation.Theme;
 import roomescape.domain.reservation.ThemeRepository;
+import roomescape.fixture.MemberFixture;
+import roomescape.fixture.ReservationFixture;
 
 @ServiceTest
 @Import(ReservationFixture.class)
